refactor(day-diet): split monolithic dayDietStore into granular stores
