+++ src/modules/diet/day-diet/application/usecases/dayChange.ts
@@ -1,22 +1,24 @@
-import { dayDietStore } from '~/modules/diet/day-diet/infrastructure/signals/dayDietStore'
+import { dayCacheStore } from '~/modules/diet/day-diet/infrastructure/signals/dayCacheStore'
+import { dayChangeStore } from '~/modules/diet/day-diet/infrastructure/signals/dayChangeStore'
+import { dayStateStore } from '~/modules/diet/day-diet/infrastructure/signals/dayStateStore'
 
-export const dayChangeData = dayDietStore.dayChangeData
+export const dayChangeData = dayChangeStore.dayChangeData
 
 /**
  * Dismisses the day change confirmation modal
  */
 export function dismissDayChangeModal() {
-  dayDietStore.setDayChangeData(null)
+  dayChangeStore.setDayChangeData(null)
 }
 
 /**
  * Accepts the day change and navigates to the new day
  */
 export function acceptDayChange() {
-  const changeData = dayDietStore.dayChangeData()
+  const changeData = dayChangeStore.dayChangeData()
   if (changeData) {
-    dayDietStore.setDayDiets([])
-    dayDietStore.setTargetDay(changeData.newDay)
-    dayDietStore.setDayChangeData(null)
+    dayCacheStore.clearCache()
+    dayStateStore.setTargetDay(changeData.newDay)
+    dayChangeStore.setDayChangeData(null)
   }
 }
